JpaRepository : fonction findOne(int id) -> findOne(long id)
service member.accession pour adherentVisualisation
service media.accession pour mediaVisualisation
service loan.ajout pour ajouter Emprunt pour adherentVisualisation et
mediaVisualisation

+++ src/main/java/org/dta/alexendrie/core/JpaRepository.java
@@ -7,7 +7,7 @@ public interface JpaRepository<T extends Model> {
 	
 	public Session getSession();
     public T save(T entity);
-    public T findOne(int id);
+    public T findOne(long id);
     public List<T> findAll();
     public void delete(T entity);
     public void delete(int id);
